+++ CHANGELOG.md
@@ -5,6 +5,22 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+#### Makefile
+
+- Changed localnet rippled image to `develop`
+- Exposed RPC port in localnet command
+
+### Removed
+
+#### xrpl/transaction
+
+- Removed integration tests for obsolete transactions `Batch` and `DelegateSet`
+
+
 ## [v.0.1.17]
 
 ### Fixed